backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
